Disable signing capabilities for ddoc, asics and scs
* MOPPIOS-625
